+++ src/main/java/org/moneymatters/mm_backend/models/dto/TransactionDTO.java
@@ -39,7 +39,7 @@ public class TransactionDTO {
         this.tagId = tagId;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
